pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "data-scribe"
version = "0.1.1"
description = "AI-powered data documentation generator for databases and dbt projects."
readme = "README.md"
authors = [
    { name = "DongWon Moon", email = "ansehddnjs187@gmail.com" },
]
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Database",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Documentation",
]

dependencies = [
    "typer[all]",
    "openai",
    "python-dotenv",
    "PyYAML",
    "ruamel.yaml",
]

[project.optional-dependencies]
postgres = ["psycopg2-binary"]
mariadb = ["mysql-connector-python"]
google = ["google-generativeai"]
ollama = ["ollama"]
confluence = ["atlassian-python-api"]
snowflake = ["snowflake-connector-python"]
bigquery = ["google-cloud-bigquery", "google-auth"]

test = [
    "pytest",
    "pytest-mock",
    "dbt-core",
    "dbt-duckdb",
]

all = [
    "psycopg2-binary",
    "mysql-connector-python",
    "google-generativeai",
    "ollama",
    "atlassian-python-api",
    "snowflake-connector-python",
    "google-cloud-bigquery",
    "google-auth",
    "pytest",
    "pytest-mock",
    "dbt-core",
    "dbt-duckdb",
]

[tool.setuptools.packages.find]
include = ["data_scribe*"]
exclude = ["jaffle_shop*", "tests*"]

[project.scripts]
data-scribe = "data_scribe.main:app"